+++ Examples/Extensions/Vector3/CubicBezierExample.cs
@@ -10,12 +10,7 @@ namespace UniSharper.Examples.Extensions.Vector3
         [SerializeField]
         private LineRenderer lineRenderer = null;
 
-        private void Update()
-        {
-            DrawCubicBezierCurve();
-        }
-
-        private void DrawCubicBezierCurve()
+        private void Start()
         {
             var points = Vector3Utility.GenerateCubicBezierPoints(UnityEngine.Vector3.zero, new UnityEngine.Vector3(0f, 0.5f), 
                 new UnityEngine.Vector3(0.75f, 0.5f, 0), new UnityEngine.Vector3(1, 0, 0), 50);