+++ CHANGELOG.md
@@ -1,9 +1,18 @@
+## [1.6.0] - 2020-05-21
+
+### Added
+
+- Adds **Vector3Utility.GenerateAchimedeanSpiralPoints** to generate points of Achimedean Spiral.
+- Adds **Vector3Utility.GenerateLogarithmicSpiralPoints** to generate points of Logarithmic Spiral.
+
+
+
 ## [1.5.0] - 2020-05-19
 
 ### Added
 
-- Adds **Vector3Utility.GenerateQuadraticBezierPoints** to generate quadratic bezier points.
-- Adds **Vector3Utility.GenerateCubicBezierPoints** to generate cubic bezier points.
+- Adds **Vector3Utility.GenerateQuadraticBezierPoints** to generate points of quadratic bezier.
+- Adds **Vector3Utility.GenerateCubicBezierPoints** to generate points of cubic bezier.